tests/wpt/web-platform-tests/css/css-text/parsing/text-align-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-align with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-align">
<meta name="assert" content="text-align supports the full grammar 'start | end | left | right | center | justify | match-parent | justify-all'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-align", "start");
test_valid_value("text-align", "end");
test_valid_value("text-align", "left");
test_valid_value("text-align", "right");
test_valid_value("text-align", "center");
test_valid_value("text-align", "justify");
test_valid_value("text-align", "match-parent");
test_valid_value("text-align", "justify-all");
</script>
</body>
</html>
